+++ src/print_tasks/B21V1PrintTask.ts
@@ -30,7 +30,8 @@ export class B21V1PrintTask extends AbstractPrintTask {
           }),
           PacketGenerator.pageEnd(),
         ],
-        this.printOptions.pageTimeoutMs
+        this.printOptions.pageTimeoutMs,
+      this.makePacketProgressCallback(),
       );
     }
   }