+++ src/print_tasks/OldD11PrintTask.ts
@@ -26,7 +26,8 @@ export class OldD11PrintTask extends AbstractPrintTask {
         ...PacketGenerator.writeImageData(image, { printheadPixels: this.printheadPixels() }),
         PacketGenerator.pageEnd(),
       ],
-      this.printOptions.pageTimeoutMs
+      this.printOptions.pageTimeoutMs,
+      this.makePacketProgressCallback(),
     );
   }