+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - The client now disconnects automatically when the number of failed heartbeats exceeds the maximum allowed value.
   - `RfidInfo` is fetched automatically when connecting and whenever the RFID state changes in a heartbeat.
 * **BREAKING:** Remove `labelType` field from `PrinterInfo` (шt's better to retrieve this property on demand).
+* `printPage` in print tasks now fires `printpacketprogress` events.
 
 # 0.46.0